% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_zcontact_dropdown.R
\name{contact_item}
\alias{contact_item}
\title{Contact Item}
\usage{
contact_item(
  name = "First Name, Last Name",
  role = "Role",
  phone = "###-###-####",
  email = "first.last@oliverwyman.com",
  logo = ""
)
}
\arguments{
\item{name}{Name}

\item{role}{Role}

\item{phone}{Phone}

\item{email}{Email}

\item{logo}{src do logo exibido como avatar redondo no topo do box
(vazio nao renderiza imagem)}
}
\value{
contact menu item
}
\description{
Creates an item to be placed in a contact dropdownmenu.
}
