% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{contatos_header}
\alias{contatos_header}
\title{Contatos do dropdown do header (usa beep_contatos)}
\usage{
contatos_header()
}
\description{
Contatos do dropdown do header (usa beep_contatos)
}
\keyword{internal}
