% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controle_execucao.R
\name{controle_preparar}
\alias{controle_preparar}
\title{Cria (se ausentes) as tabelas de controle e versoes de carga no beepdb}
\usage{
controle_preparar()
}
\description{
Cria (se ausentes) as tabelas de controle e versoes de carga no beepdb
}
