% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifica_fonte.R
\name{.artefato_cache}
\alias{.artefato_cache}
\title{Caminho do 1o artefato saveRDS() do script (relativo a raiz), ou NULL}
\usage{
.artefato_cache(exprs)
}
\description{
Caminho do 1o artefato saveRDS() do script (relativo a raiz), ou NULL
}
\keyword{internal}
