% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifica_fonte.R
\docType{data}
\name{.assunto_educabr}
\alias{.assunto_educabr}
\title{Mapa funcao edubr (ex-educabR) -> filtro de assunto no metainep (sem acentos)}
\format{
An object of class \code{character} of length 3.
}
\usage{
.assunto_educabr
}
\description{
Mapa funcao edubr (ex-educabR) -> filtro de assunto no metainep (sem acentos)
}
\keyword{internal}
