% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorporar_municipio_ibge.R
\name{.baixar_json_ibge}
\alias{.baixar_json_ibge}
\title{Baixa e le um JSON da API do IBGE com tratamento de erro}
\usage{
.baixar_json_ibge(url)
}
\description{
Baixa e le um JSON da API do IBGE com tratamento de erro
}
\keyword{internal}
