% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incorporar_municipio_ibge.R
\name{.grupo_por_nome}
\alias{.grupo_por_nome}
\title{datagroup_id pelo nome dentro de uma familia (parent)}
\usage{
.grupo_por_nome(con, nome, parent_id)
}
\description{
datagroup_id pelo nome dentro de uma familia (parent)
}
\keyword{internal}
