% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executa_atualizacao.R
\name{.heads_bare}
\alias{.heads_bare}
\title{Nomes de funcao chamados sem namespace no script (so o parse)}
\usage{
.heads_bare(exprs)
}
\description{
Nomes de funcao chamados sem namespace no script (so o parse)
}
\keyword{internal}
