% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifica_fonte.R
\docType{data}
\name{.leitores_url}
\alias{.leitores_url}
\title{Leitores/downloaders cujo 1o argumento string http(s) caracteriza coleta}
\format{
An object of class \code{character} of length 11.
}
\usage{
.leitores_url
}
\description{
Leitores/downloaders cujo 1o argumento string http(s) caracteriza coleta
}
\keyword{internal}
