% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencias.R
\name{.ler_manifesto_arquivo}
\alias{.ler_manifesto_arquivo}
\title{Le e valida o manifesto CSV de dependencias}
\usage{
.ler_manifesto_arquivo(arquivo)
}
\description{
Le e valida o manifesto CSV de dependencias
}
\keyword{internal}
