% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifica_fonte.R
\name{.metainep_edubr}
\alias{.metainep_edubr}
\title{Catalogo metainep do edubr (dataset lazy-data, nao vai em exports)}
\usage{
.metainep_edubr()
}
\description{
Catalogo metainep do edubr (dataset lazy-data, nao vai em exports)
}
\keyword{internal}
