% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencias.R
\docType{data}
\name{.reguas_dependencia}
\alias{.reguas_dependencia}
\title{Valores validos do manifesto (fase 1)}
\format{
An object of class \code{character} of length 2.
}
\usage{
.reguas_dependencia
}
\description{
Valores validos do manifesto (fase 1)
}
\keyword{internal}
