% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencias.R
\name{.series_proprias_manifesto}
\alias{.series_proprias_manifesto}
\title{Union das series proprias declaradas nas linhas do manifesto do script
(coluna serie_propria, ";"-separada)}
\usage{
.series_proprias_manifesto(linhas)
}
\description{
Union das series proprias declaradas nas linhas do manifesto do script
(coluna serie_propria, ";"-separada)
}
\keyword{internal}
