% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controle_execucao.R
\name{espelhar_dumps}
\alias{espelhar_dumps}
\title{Espelha os dumps mantidos em diretorio secundario (extrainters quando montado)}
\usage{
espelhar_dumps(dir_dump = "~/backups_beepdb", espelho)
}
\description{
Espelha os dumps mantidos em diretorio secundario (extrainters quando montado)
}
\keyword{internal}
