% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_indicadores.R
\name{explorar_badge_n}
\alias{explorar_badge_n}
\title{Texto do badge de n de pares: NA quando o n basta; abaixo do minimo
matematico (3) a correlacao nem existe}
\usage{
explorar_badge_n(n, minimo = explorar_n_min)
}
\description{
Texto do badge de n de pares: NA quando o n basta; abaixo do minimo
matematico (3) a correlacao nem existe
}
\keyword{internal}
