% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_indicadores.R
\name{explorar_cor}
\alias{explorar_cor}
\title{Correlacoes simples de um par (Pearson e Spearman) com n de pares;
series degeneradas (dp~0) ou com menos de 3 pares devolvem NA com
aviso explicativo}
\usage{
explorar_cor(pares)
}
\description{
Correlacoes simples de um par (Pearson e Spearman) com n de pares;
series degeneradas (dp~0) ou com menos de 3 pares devolvem NA com
aviso explicativo
}
\keyword{internal}
