% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_indicadores.R
\name{explorar_dicionario}
\alias{explorar_dicionario}
\title{Dicionario do catalogo: mdata + classe, frequencia, unidades, fonte e
observacoes (joins fail-open: DW sem auxiliares segue com colunas NA)}
\usage{
explorar_dicionario(con)
}
\description{
Dicionario do catalogo: mdata + classe, frequencia, unidades, fonte e
observacoes (joins fail-open: DW sem auxiliares segue com colunas NA)
}
\keyword{internal}
