% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_indicadores.R
\docType{data}
\name{explorar_n_min}
\alias{explorar_n_min}
\title{Limiar do badge visual de "n insuficiente" (nada e escondido: o n
aparece sempre junto ao r)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
explorar_n_min
}
\description{
Limiar do badge visual de "n insuficiente" (nada e escondido: o n
aparece sempre junto ao r)
}
\keyword{datasets}
