% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_indicadores.R
\name{explorar_nomes_locais}
\alias{explorar_nomes_locais}
\title{Nomes das localidades indexados pelo local_id (hover dos graficos)}
\usage{
explorar_nomes_locais(con)
}
\description{
Nomes das localidades indexados pelo local_id (hover dos graficos)
}
\keyword{internal}
