% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controle_execucao.R
\name{fingerprint_beepdb}
\alias{fingerprint_beepdb}
\title{Fingerprint barato do conteudo do beepdb: volumetria + soma + ultimas
escritas. Muda com qualquer INSERT/UPDATE/replace de indicador.}
\usage{
fingerprint_beepdb()
}
\description{
Fingerprint barato do conteudo do beepdb: volumetria + soma + ultimas
escritas. Muda com qualquer INSERT/UPDATE/replace de indicador.
}
\keyword{internal}
