% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravar_serie_dw.R
\name{gravar_serie_dw}
\alias{gravar_serie_dw}
\title{Grava (recalculando por completo ou acrescentando) a serie de um indicador}
\usage{
gravar_serie_dw(orig_name, serie, modo = c("replace", "append"))
}
\arguments{
\item{orig_name}{nome do indicador em mdata}

\item{serie}{data.frame com \code{local}, \code{periodo}, \code{valor}}

\item{modo}{\code{"replace"} (default) ou \code{"append"}}
}
\description{
Grava (recalculando por completo ou acrescentando) a serie de um indicador
}
