% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{logo_header_tag}
\alias{logo_header_tag}
\title{Logo do header (usa beep_logo e beep_logo_width)}
\usage{
logo_header_tag()
}
\description{
Logo do header (usa beep_logo e beep_logo_width)
}
\keyword{internal}
