% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{logo_rodape_tag}
\alias{logo_rodape_tag}
\title{Logo+link do rodape da sidebar (usa beep_logo e beep_logo_link)}
\usage{
logo_rodape_tag(width = 200)
}
\description{
Logo+link do rodape da sidebar (usa beep_logo e beep_logo_link)
}
\keyword{internal}
