% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explorar_bivar.R
\name{mod_explorar_bivar_ui}
\alias{mod_explorar_bivar_ui}
\title{UI da correlacao par a par (aba "Análise Bi-Variada")}
\usage{
mod_explorar_bivar_ui(id)
}
\description{
UI da correlacao par a par (aba "Análise Bi-Variada")
}
\keyword{internal}
