% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explorar_dicionario.R
\name{mod_explorar_dicionario_server}
\alias{mod_explorar_dicionario_server}
\title{Server do dicionario: catalogo completo com filtros e export}
\usage{
mod_explorar_dicionario_server(id)
}
\description{
Server do dicionario: catalogo completo com filtros e export
}
\keyword{internal}
