% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explorar_dicionario.R
\name{mod_explorar_dicionario_ui}
\alias{mod_explorar_dicionario_ui}
\title{UI do dicionario de dados do catalogo (aba "Dicionário de Dados")}
\usage{
mod_explorar_dicionario_ui(id)
}
\description{
UI do dicionario de dados do catalogo (aba "Dicionário de Dados")
}
\keyword{internal}
