% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explorar_dist.R
\name{mod_explorar_dist_server}
\alias{mod_explorar_dist_server}
\title{Server das distribuicoes: histograma do ano escolhido e boxplot da
serie toda por ano, sobre a serie achatada}
\usage{
mod_explorar_dist_server(id)
}
\description{
Server das distribuicoes: histograma do ano escolhido e boxplot da
serie toda por ano, sobre a serie achatada
}
\keyword{internal}
