% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explorar_dist.R
\name{mod_explorar_dist_ui}
\alias{mod_explorar_dist_ui}
\title{UI das distribuicoes do indicador (aba "Distribuições")}
\usage{
mod_explorar_dist_ui(id)
}
\description{
UI das distribuicoes do indicador (aba "Distribuições")
}
\keyword{internal}
