% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explorar_matriz.R
\name{mod_explorar_matriz_ui}
\alias{mod_explorar_matriz_ui}
\title{UI da matriz de correlacoes (aba "Matriz de correlação", F3)}
\usage{
mod_explorar_matriz_ui(id)
}
\description{
UI da matriz de correlacoes (aba "Matriz de correlação", F3)
}
\keyword{internal}
