% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explorar_preditores.R
\name{mod_explorar_preditores_ui}
\alias{mod_explorar_preditores_ui}
\title{UI do ranking de preditores (aba "Relações preditoras", F4)}
\usage{
mod_explorar_preditores_ui(id)
}
\description{
UI do ranking de preditores (aba "Relações preditoras", F4)
}
\keyword{internal}
