% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explorar_univar.R
\name{mod_explorar_univar_server}
\alias{mod_explorar_univar_server}
\title{Server da ficha: cobertura e dispersao por ano (numeros da heuristica
de banda/lolly do roadmap) lidas da serie achatada}
\usage{
mod_explorar_univar_server(id)
}
\description{
Server da ficha: cobertura e dispersao por ano (numeros da heuristica
de banda/lolly do roadmap) lidas da serie achatada
}
\keyword{internal}
