% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explorar_univar.R
\name{mod_explorar_univar_ui}
\alias{mod_explorar_univar_ui}
\title{UI da ficha do indicador (aba "Análise Univariada")}
\usage{
mod_explorar_univar_ui(id)
}
\description{
UI da ficha do indicador (aba "Análise Univariada")
}
\keyword{internal}
