% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_ui.R
\name{painel_abas}
\alias{painel_abas}
\title{Abas do painel (uma tabPanel por modulo) — solte/adicione a vontade}
\usage{
painel_abas()
}
\description{
Abas do painel (uma tabPanel por modulo) — solte/adicione a vontade
}
\keyword{internal}
