% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_anos}
\alias{painel_anos}
\title{Anos com observacoes de um indicador (para os limites do slider do mapa)}
\usage{
painel_anos(con, mdata_id)
}
\description{
Anos com observacoes de um indicador (para os limites do slider do mapa)
}
\keyword{internal}
