% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_anos_cache}
\alias{painel_anos_cache}
\title{Anos com observacoes de um indicador, cacheado}
\usage{
painel_anos_cache(mdata_id)
}
\description{
Anos com observacoes de um indicador, cacheado
}
\keyword{internal}
