% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_ui.R
\name{painel_aviso_carregando}
\alias{painel_aviso_carregando}
\title{Aviso de carregamento (CSS mostra/esconde conforme a conexao)}
\usage{
painel_aviso_carregando()
}
\description{
Aviso de carregamento (CSS mostra/esconde conforme a conexao)
}
\keyword{internal}
