% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_branding.R
\name{painel_brand_apoio_box}
\alias{painel_brand_apoio_box}
\title{Um box de apoio da aba Sobre: logo clicavel ao lado da frase}
\usage{
painel_brand_apoio_box(logo, url, frase, nome)
}
\description{
Um box de apoio da aba Sobre: logo clicavel ao lado da frase
}
\keyword{internal}
