% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_branding.R
\name{painel_brand_paleta}
\alias{painel_brand_paleta}
\title{Paleta inicial do painel (usa painel_paleta)}
\usage{
painel_brand_paleta(default = "govbr")
}
\description{
Paleta inicial do painel (usa painel_paleta)
}
\keyword{internal}
