% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_branding.R
\name{painel_brand_titulo}
\alias{painel_brand_titulo}
\title{Titulo do painel (usa painel_titulo)}
\usage{
painel_brand_titulo(default = "Painel de Indicadores")
}
\description{
Titulo do painel (usa painel_titulo)
}
\keyword{internal}
