% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_cache.R
\name{painel_cache_chave}
\alias{painel_cache_chave}
\title{Chave de cache com host+dbname sanitizados na frente}
\usage{
painel_cache_chave(sufixo)
}
\description{
Chave de cache com host+dbname sanitizados na frente
}
\keyword{internal}
