% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_cache.R
\name{painel_cache_dir}
\alias{painel_cache_dir}
\title{Diretorio do cache em disco, relativo ao diretorio do app (como www/)}
\usage{
painel_cache_dir()
}
\description{
Diretorio do cache em disco, relativo ao diretorio do app (como www/)
}
\keyword{internal}
