% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_cache.R
\name{painel_cache_limpar}
\alias{painel_cache_limpar}
\title{Descarta o cache de memoria e de disco (rodar apos atualizar o DW com ETL)}
\usage{
painel_cache_limpar()
}
\description{
Descarta o cache de memoria e de disco (rodar apos atualizar o DW com ETL)
}
\keyword{internal}
