% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_codigo_mun_cache}
\alias{painel_codigo_mun_cache}
\title{Codigo IBGE dos municipios, cacheado}
\usage{
painel_codigo_mun_cache()
}
\description{
Codigo IBGE dos municipios, cacheado
}
\keyword{internal}
