% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_compostos}
\alias{painel_compostos}
\title{Indicadores compostos do catalogo (data_class_id = 4 no mdata_exts)}
\usage{
painel_compostos(con)
}
\description{
Indicadores compostos do catalogo (data_class_id = 4 no mdata_exts)
}
\keyword{internal}
