% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_compostos_cache}
\alias{painel_compostos_cache}
\title{Indicadores compostos do catalogo, cacheado}
\usage{
painel_compostos_cache()
}
\description{
Indicadores compostos do catalogo, cacheado
}
\keyword{internal}
