% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_filtrar_ano}
\alias{painel_filtrar_ano}
\title{Fatiador puro do cache achatado: devolve as linhas do ano com as colunas
do contrato historico (sem a coluna ano), vazio tipado quando nao ha match}
\usage{
painel_filtrar_ano(d, ano)
}
\description{
Fatiador puro do cache achatado: devolve as linhas do ano com as colunas
do contrato historico (sem a coluna ano), vazio tipado quando nao ha match
}
\keyword{internal}
