% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_geo_local_cache}
\alias{painel_geo_local_cache}
\title{Geometria de uma localidade, cacheada}
\usage{
painel_geo_local_cache(local_id)
}
\description{
Geometria de uma localidade, cacheada
}
\keyword{internal}
