% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_geo_nivel_cache}
\alias{painel_geo_nivel_cache}
\title{Geometrias de um nivel territorial, cacheado}
\usage{
painel_geo_nivel_cache(nivel_id)
}
\description{
Geometrias de um nivel territorial, cacheado
}
\keyword{internal}
