% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_geo_pai_uf}
\alias{painel_geo_pai_uf}
\title{Geometria da UF que contem uma localidade (contexto do globo: o
estado aparece inteiro com a localidade em destaque)}
\usage{
painel_geo_pai_uf(con, local_id)
}
\description{
Geometria da UF que contem uma localidade (contexto do globo: o
estado aparece inteiro com a localidade em destaque)
}
\keyword{internal}
