% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_geo_pai_uf_cache}
\alias{painel_geo_pai_uf_cache}
\title{Geometria da UF de uma localidade, cacheada}
\usage{
painel_geo_pai_uf_cache(local_id)
}
\description{
Geometria da UF de uma localidade, cacheada
}
\keyword{internal}
