% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_geo_uf}
\alias{painel_geo_uf}
\title{Geometrias das Unidades da Federacao do DW (largura 2 do geoloc_id:
26 UFs + DF), com code/label para o globo}
\usage{
painel_geo_uf(con)
}
\description{
Geometrias das Unidades da Federacao do DW (largura 2 do geoloc_id:
26 UFs + DF), com code/label para o globo
}
\keyword{internal}
