% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_geo_vazio}
\alias{painel_geo_vazio}
\title{sf vazio tipado (colunas code/label/geometry) para as leituras de
geometria do globo}
\usage{
painel_geo_vazio()
}
\description{
sf vazio tipado (colunas code/label/geometry) para as leituras de
geometria do globo
}
\keyword{internal}
