% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_geojson}
\alias{painel_geojson}
\title{GeoJSON (string) de um objeto sf para mensagens Shiny ao cliente}
\usage{
painel_geojson(x)
}
\description{
GeoJSON (string) de um objeto sf para mensagens Shiny ao cliente
}
\keyword{internal}
