% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\docType{data}
\name{painel_grupos_selecao}
\alias{painel_grupos_selecao}
\title{Ordem canonica dos grupos no seletor de indicadores do painel}
\format{
An object of class \code{character} of length 12.
}
\usage{
painel_grupos_selecao
}
\description{
Ordem canonica dos grupos no seletor de indicadores do painel
}
\keyword{internal}
