% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_hierarquia_cache}
\alias{painel_hierarquia_cache}
\title{Hierarquia de agrupamentos (Eixos/Objetivos), cacheado}
\usage{
painel_hierarquia_cache()
}
\description{
Hierarquia de agrupamentos (Eixos/Objetivos), cacheado
}
\keyword{internal}
