% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_locais_com_dados_cache}
\alias{painel_locais_com_dados_cache}
\title{Localidades de um nivel com dados para um indicador, cacheado}
\usage{
painel_locais_com_dados_cache(mdata_id, nivel_id)
}
\description{
Localidades de um nivel com dados para um indicador, cacheado
}
\keyword{internal}
