% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_locais_nivel_cache}
\alias{painel_locais_nivel_cache}
\title{Localidades de um nivel territorial, cacheado}
\usage{
painel_locais_nivel_cache(nivel_id)
}
\description{
Localidades de um nivel territorial, cacheado
}
\keyword{internal}
