% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_local_top_cache}
\alias{painel_local_top_cache}
\title{Localidade com maior cobertura de um indicador num nivel, cacheado}
\usage{
painel_local_top_cache(mdata_id, nivel_id)
}
\description{
Localidade com maior cobertura de um indicador num nivel, cacheado
}
\keyword{internal}
