% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_local_top_uf_cache}
\alias{painel_local_top_uf_cache}
\title{Localidade com maior cobertura numa UF, cacheada}
\usage{
painel_local_top_uf_cache(mdata_id, nivel_id, uf_local_id)
}
\description{
Localidade com maior cobertura numa UF, cacheada
}
\keyword{internal}
