% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_municipio_filtro}
\alias{painel_municipio_filtro}
\title{Fragmento SQL que seleciona apenas municipios (alias \code{l} no chamador)}
\usage{
painel_municipio_filtro(alias = "l")
}
\description{
Fragmento SQL que seleciona apenas municipios (alias \code{l} no chamador)
}
