% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_niveis_cache}
\alias{painel_niveis_cache}
\title{Niveis territoriais disponiveis, cacheado}
\usage{
painel_niveis_cache()
}
\description{
Niveis territoriais disponiveis, cacheado
}
\keyword{internal}
