% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_nivel_default}
\alias{painel_nivel_default}
\title{Nivel territorial de abertura default da aba Regiao: municipal quando
disponivel, senao UF}
\usage{
painel_nivel_default(niveis)
}
\description{
Nivel territorial de abertura default da aba Regiao: municipal quando
disponivel, senao UF
}
\keyword{internal}
