% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_num}
\alias{painel_num}
\title{Numeros no padrao pt-BR para rotulos de mapa e legenda}
\usage{
painel_num(x)
}
\description{
Numeros no padrao pt-BR para rotulos de mapa e legenda
}
\keyword{internal}
