% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_ranking_local_cache}
\alias{painel_ranking_local_cache}
\title{Ranking de uma localidade num indicador e ano, cacheado}
\usage{
painel_ranking_local_cache(mdata_id, local_id, ano)
}
\description{
Ranking de uma localidade num indicador e ano, cacheado
}
\keyword{internal}
