% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_ranking_vazio}
\alias{painel_ranking_vazio}
\title{Ranking vazio (localidade nao municipal, sem dado no ano ou sem par)}
\usage{
painel_ranking_vazio()
}
\description{
Ranking vazio (localidade nao municipal, sem dado no ano ou sem par)
}
\keyword{internal}
