% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_ufs_com_dados_cache}
\alias{painel_ufs_com_dados_cache}
\title{UFs com dados num nivel, cacheado}
\usage{
painel_ufs_com_dados_cache(mdata_id, nivel_id)
}
\description{
UFs com dados num nivel, cacheado
}
\keyword{internal}
