% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_valores_local_cache}
\alias{painel_valores_local_cache}
\title{Serie de um indicador em uma localidade, cacheado}
\usage{
painel_valores_local_cache(mdata_id, local_id)
}
\description{
Serie de um indicador em uma localidade, cacheado
}
\keyword{internal}
