% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_dw.R
\name{painel_valores_local_todos_cache}
\alias{painel_valores_local_todos_cache}
\title{Valores de todos os indicadores em uma localidade, cacheado}
\usage{
painel_valores_local_todos_cache(local_id)
}
\description{
Valores de todos os indicadores em uma localidade, cacheado
}
\keyword{internal}
