% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_panel.R
\name{panel_server}
\alias{panel_server}
\title{Server do painel (um mod_*_server por aba)}
\usage{
panel_server(input, output, session)
}
\description{
Server do painel (um mod_*_server por aba)
}
\keyword{internal}
