% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_panel_sobre.R
\name{panel_sobre_server}
\alias{panel_sobre_server}
\title{Acoes da aba Sobre (navegacao para as demais abas) — chamada no server
do panel_app()}
\usage{
panel_sobre_server(input, session)
}
\description{
Acoes da aba Sobre (navegacao para as demais abas) — chamada no server
do panel_app()
}
\keyword{internal}
