% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{resolver_logo_contato}
\alias{resolver_logo_contato}
\title{Logo de um box de contato: 5o campo da entrada, beep_contato_logo ou o
padrao quadrado da marca; "none" (ou "-") desliga}
\usage{
resolver_logo_contato(campo = "")
}
\description{
Logo de um box de contato: 5o campo da entrada, beep_contato_logo ou o
padrao quadrado da marca; "none" (ou "-") desliga
}
\keyword{internal}
