% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branding.R
\name{resolver_marca_src}
\alias{resolver_marca_src}
\title{Resolve o src de uma marca (URL, arquivo local ou www/ do pacote)}
\usage{
resolver_marca_src(marca)
}
\description{
Resolve o src de uma marca (URL, arquivo local ou www/ do pacote)
}
\keyword{internal}
