% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beep-package.R
\docType{package}
\name{beep-package}
\alias{beep}
\alias{beep-package}
\title{beep: Pacote de inclusão de fontes de dados, indicadores, e filtros de variáveis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

AED interativa e adição de dados. beep é o acrônimo de
'Back-End de Exploração e Painéis de dados'. Sucessor do pacote AEDi
(https://github.com/DistintiveLab/AEDi).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DistintiveLab/beep}
  \item Report bugs at \url{https://github.com/DistintiveLab/beep/issues}
}

}
\author{
\strong{Maintainer}: Rodrigo Borges \email{rodrigo@borges.net.br}

}
\keyword{internal}
