% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_ui.R
\name{painel_recursos}
\alias{painel_recursos}
\title{Recursos de cabecalho (meta, CSS/JS do painel, Gov.br) e div raiz da paleta}
\usage{
painel_recursos(assets_dir, paleta = c("govbr", "pb"))
}
\description{
Recursos de cabecalho (meta, CSS/JS do painel, Gov.br) e div raiz da paleta
}
\keyword{internal}
